refactor: hoist formatLastContact to src/lib for reuse
Move the relative-time helper from the pipeline _lib directory to
src/lib/format-relative-time so both the pipeline and the new
conversation history UI can import it without cross-feature _lib
imports. Add characterization tests in the new location and delete
the now-redundant pipeline copy.

Refs: #131
